chore: update imports after module renaming
