+++ analyzers/ngrams/ngrams_stats/__init__.py
@@ -3,4 +3,4 @@ from analyzer_interface import SecondaryAnalyzerDeclaration
 from .interface import interface
 from .main import main
 
-ngram_stats = SecondaryAnalyzerDeclaration(interface=interface, main=main)
+ngrams_stats = SecondaryAnalyzerDeclaration(interface=interface, main=main)
