+++ analyzers/ngrams/test_ngram_stats.py
@@ -2,13 +2,13 @@ from pathlib import Path
 
 from testing import ParquetTestData, test_secondary_analyzer
 
-from .ngram_stats.interface import OUTPUT_NGRAM_FULL, OUTPUT_NGRAM_STATS, interface
-from .ngram_stats.main import main
 from .ngrams_base.interface import (
     OUTPUT_MESSAGE,
     OUTPUT_MESSAGE_NGRAMS,
     OUTPUT_NGRAM_DEFS,
 )
+from .ngrams_stats.interface import OUTPUT_NGRAM_FULL, OUTPUT_NGRAM_STATS, interface
+from .ngrams_stats.main import main
 from .test_data import test_data_dir
